data_engineering/ingestion/requirements.txt
# Data Engineering - Ingestion Dependencies

# Kafka/Redpanda client
confluent-kafka>=2.3.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# HTTP clients for API data sources
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# PostgreSQL
psycopg2-binary>=2.9.9

# Environment management
python-dotenv>=1.0.0

# Logging and utilities
loguru>=0.7.0
tenacity>=8.2.0  # Retry logic

# Type hints
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
